+++ src/calendarjs.js
@@ -1,5 +1,5 @@
 /*
- * Calendar.js Library v1.4.1
+ * Calendar.js Library v1.4.2
  *
  * Copyright 2021 Bunoon
  * Released under the GNU AGPLv3 license
@@ -221,6 +221,7 @@
  * @property    {string}    notSearchText                               The text that should be displayed for the "Not (opposite)" label.
  * @property    {string}    showEmptyDaysInWeekViewText                 The text that should be displayed for the "Show empty days in the week view" label.
  * @property    {string}    showHolidaysInTheDisplaysText               The text that should be displayed for the "Show holidays in the main display and title bars" label.
+ * @property    {string}    newEventDefaultTitle                        The default title that should be used for new events (defaults to "* New Event").
  */
 
 
@@ -267,6 +268,7 @@
  * @property    {boolean}   showEmptyDaysInWeekView                     States if empty days should be shown in the Week view (defaults to true).
  * @property    {boolean}   hideEventsWithoutGroupAssigned              States if events without a group should be hidden (defaults to false).
  * @property    {boolean}   showHolidays                                States if the holidays should be shown (defaults to true).
+ * @property    {boolean}   useTemplateWhenAddingNewEvent               States if a blank template event should be added when adding a new event (causing the dialog to be in edit mode, defaults to true).
  */
 
 
@@ -1620,7 +1622,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                             event.appendChild( icon );
                         }
     
-                        event.innerHTML += eventTitle;
+                        event.innerHTML += stripHTMLTagsFromText( eventTitle );
     
                     } else {
                         event.className += " event-circle";
@@ -1834,7 +1836,15 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         buildToolbarButton( titleBar, "ib-close", _options.closeTooltipText, hideFullDayView );
 
         if ( _options.manualEditingEnabled ) {
-            buildToolbarButton( titleBar, "ib-plus", _options.addEventTooltipText, addNewEvent );
+            buildToolbarButton( titleBar, "ib-plus", _options.addEventTooltipText, function() {
+                if ( _options.useTemplateWhenAddingNewEvent ) {
+                    var newBlankTemplateEvent = buildBlankTemplateEvent( _element_FullDayView_DateSelected, _element_FullDayView_DateSelected );
+    
+                    showEventEditingDialog( newBlankTemplateEvent );
+                } else {
+                    addNewEvent();
+                }
+            } );
         }
 
         buildToolbarButton( titleBar, "ib-arrow-right-full", _options.nextDayTooltipText, onNextDay );
@@ -1896,7 +1906,14 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 minutesFromTop = Math.floor( ( e.pageY - ( contentHoursOffset.top + scrollPosition.top ) ) / pixelsPerMinute ),
                 hoursMinutes = getHoursAndMinutesFromMinutes( minutesFromTop );
             
-            showEventEditingDialog( null, _element_FullDayView_DateSelected, hoursMinutes );
+            if ( _options.useTemplateWhenAddingNewEvent ) {
+                var newBlankTemplateEventTime = padNumber( hoursMinutes[ 0 ] ) + ":" + padNumber( hoursMinutes[ 1 ] ),
+                    newBlankTemplateEvent = buildBlankTemplateEvent( _element_FullDayView_DateSelected, _element_FullDayView_DateSelected, newBlankTemplateEventTime, newBlankTemplateEventTime );
+
+                showEventEditingDialog( newBlankTemplateEvent );
+            } else {
+                showEventEditingDialog( null, _element_FullDayView_DateSelected, hoursMinutes );
+            }
         }
     }
 
@@ -2074,7 +2091,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 title.appendChild( icon );
             }
             
-            title.innerHTML += eventDetails.title;
+            title.innerHTML += stripHTMLTagsFromText( eventDetails.title );
             event.appendChild( title );
     
             if ( !eventDetails.isAllDay || _options.showAllDayEventDetailsInFullDayView ) {
@@ -2458,7 +2475,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 title.appendChild( icon );
             }
             
-            title.innerHTML += eventDetails.title;
+            title.innerHTML += stripHTMLTagsFromText( eventDetails.title );
             event.appendChild( title );
     
             var startTime = createElement( "div", "date" );
@@ -2537,7 +2554,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 var addNewEventDate = new Date( date.getFullYear(), date.getMonth(), 1 );
 
                 buildToolbarButton( header, "ib-plus", _options.addEventTooltipText, function() {
-                    showEventEditingDialog( null, addNewEventDate );
+                    if ( _options.useTemplateWhenAddingNewEvent ) {
+                        var newBlankTemplateEvent = buildBlankTemplateEvent( addNewEventDate, addNewEventDate );
+
+                        showEventEditingDialog( newBlankTemplateEvent );
+                    } else {
+                        showEventEditingDialog( null, addNewEventDate );
+                    }
                 } );
             }
 
@@ -2810,7 +2833,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                 title.appendChild( icon );
             }
             
-            title.innerHTML += eventDetails.title;
+            title.innerHTML += stripHTMLTagsFromText( eventDetails.title );
             event.appendChild( title );
 
             var startTime = createElement( "div", "date" );
@@ -2886,7 +2909,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                     showFullDayView( expandDate, true );
                 },
                 addEventFunction = function() {
-                    showEventEditingDialog( null, expandDate );
+                    if ( _options.useTemplateWhenAddingNewEvent ) {
+                        var newBlankTemplateEvent = buildBlankTemplateEvent( expandDate, expandDate );
+
+                        showEventEditingDialog( newBlankTemplateEvent );
+                    } else {
+                        showEventEditingDialog( null, expandDate );
+                    }
                 };
 
             var day = createElement( "div", "day" );
@@ -3166,7 +3195,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
             if ( _options.manualEditingEnabled ) {
                 dayElement.ondblclick = function() {
-                    showEventEditingDialog( null, dayDate );
+                    if ( _options.useTemplateWhenAddingNewEvent ) {
+                        var newBlankTemplateEvent = buildBlankTemplateEvent( dayDate, dayDate );
+
+                        showEventEditingDialog( newBlankTemplateEvent );
+                    } else {
+                        showEventEditingDialog( null, dayDate );
+                    }
                 };
 
                 makeAreaDroppable( dayElement, year, month, actualDay );
@@ -3390,7 +3425,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         if ( _options.manualEditingEnabled ) {
             buildMenuItemWithIcon( _element_DropDownMenu_Day, "ib-plus-icon", _options.addEventTitle + "...", function() {
-                showEventEditingDialog( null, _element_DropDownMenu_Day_DateSelected );
+                if ( _options.useTemplateWhenAddingNewEvent ) {
+                    var newBlankTemplateEvent = buildBlankTemplateEvent( _element_DropDownMenu_Day_DateSelected, _element_DropDownMenu_Day_DateSelected );
+
+                    showEventEditingDialog( newBlankTemplateEvent );
+                } else {
+                    showEventEditingDialog( null, _element_DropDownMenu_Day_DateSelected );
+                }
             }, true );
     
             buildMenuSeparator( _element_DropDownMenu_Day );
@@ -3504,7 +3545,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             _document.body.appendChild( _element_DropDownMenu_FullDay );
 
             buildMenuItemWithIcon( _element_DropDownMenu_FullDay, "ib-plus-icon", _options.addEventTitle + "...", function() {
-                showEventEditingDialog( null, _element_FullDayView_DateSelected );
+                if ( _options.useTemplateWhenAddingNewEvent ) {
+                    var newBlankTemplateEvent = buildBlankTemplateEvent( _element_FullDayView_DateSelected, _element_FullDayView_DateSelected );
+
+                    showEventEditingDialog( newBlankTemplateEvent );
+                } else {
+                    showEventEditingDialog( null, _element_FullDayView_DateSelected );
+                }
             }, true );
 
             _element_DropDownMenu_FullDay_RemoveEvents_Separator = buildMenuSeparator( _element_DropDownMenu_FullDay );
@@ -3540,6 +3587,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
             triggerOptionsEventWithData( "onOptionsUpdated", _options );
             build( _currentDate, true );
         }, true );
+
+        buildMenuSeparator( _element_DropDownMenu_HeaderDay );
+
+        buildMenuItemWithIcon( _element_DropDownMenu_HeaderDay, "ib-octagon-hollow-icon", _options.visibleDaysTabText + "...", function() {
+            selectTab( _element_ConfigurationDialog, 3 );
+            showConfigurationDialog();
+        } );
     }
 
     function buildMenuItemWithIcon( container, iconCSS, text, onClickEvent, isBold ) {
@@ -3908,7 +3962,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
     function showEventEditingDialog( eventDetails, overrideTodayDate, overrideTimeValues ) {
         addNode( _document.body, _element_DisabledBackground );
-        selectFirstTab( _element_EventEditorDialog );
+        selectTab( _element_EventEditorDialog );
 
         if ( isDefined( eventDetails ) ) {
             _element_EventEditorDialog_OKButton.value = _options.updateText;
@@ -4178,6 +4232,43 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         updateFullDayViewFromEventEdit();
         updateViewAllEventsViewFromEventEdit();
         updateViewAllWeekEventsViewFromEventEdit();
+    }
+
+    function buildBlankTemplateEvent( fromDate, toDate, fromTime, toTime ) {
+        fromTime = isDefined( fromTime ) ? fromTime : "09:00";
+        toTime = isDefined( toTime ) ? fromTime : "09:00";
+
+        setTimeOnDate( fromDate, fromTime );
+        setTimeOnDate( toDate, toTime );
+
+        var newEvent = {
+            from: fromDate,
+            to: toDate,
+            title: _options.newEventDefaultTitle,
+            description: "",
+            location: "",
+            group: "",
+            isAllDay: false,
+            showAlerts: true,
+            color: _options.defaultEventBackgroundColor,
+            colorText: _options.defaultEventTextColor,
+            colorBorder: _options.defaultEventBorderColor,
+            repeatEveryExcludeDays: [],
+            repeatEnds: null,
+            url: "",
+            repeatEveryCustomValue: "",
+            repeatEvery: _repeatType.never,
+            repeatEveryCustomType: _repeatCustomType.daily,
+            organizerName: "",
+            organizerEmailAddress: "",
+        };
+
+        _this.addEvent( newEvent, false );
+
+        buildDayEvents();
+        refreshOpenedViews();
+
+        return newEvent;
     }
 
 
@@ -5217,7 +5308,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
                             _element_Tooltip_Title.appendChild( icon );
                         }
                         
-                        _element_Tooltip_Title.innerHTML += eventDetails.title;
+                        _element_Tooltip_Title.innerHTML += stripHTMLTagsFromText( eventDetails.title );
 
                         if ( isDefinedNumber( eventDetails.repeatEvery ) && eventDetails.repeatEvery > _repeatType.never ) {
                             _element_Tooltip_Repeats.innerText = _options.repeatsText.replace( ":", "" ) + " " + getRepeatsText( eventDetails.repeatEvery );
@@ -5358,12 +5449,14 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         contents.style.display = "block";
     }
 
-    function selectFirstTab( container ) {
+    function selectTab( container, tabIndex ) {
+        tabIndex = isDefined( tabIndex ) ? tabIndex : 0;
+
         var tabs = container.getElementsByClassName( "tab" ),
             tabsLength = tabs.length;
         
         if ( tabsLength > 0 ) {
-            tabs[ 0 ].click();
+            tabs[ tabIndex ].click();
         }
     }
 
@@ -5732,6 +5825,13 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
         if ( element !== null ) {
             element.className = className;
         }
+    }
+
+    function stripHTMLTagsFromText( text ) {
+        var div = createElement( "div" );
+        div.innerHTML = text;
+
+        return div.innerText;
     }
 
 
@@ -7378,7 +7478,7 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
      * @returns     {string}                                                The version number.
      */
     this.getVersion = function() {
-        return "1.4.1";
+        return "1.4.2";
     };
 
 
@@ -7625,6 +7725,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         if ( !isDefinedBoolean( _options.showHolidays ) ) {
             _options.showHolidays = true;
+        }
+
+        if ( !isDefinedBoolean( _options.useTemplateWhenAddingNewEvent ) ) {
+            _options.useTemplateWhenAddingNewEvent = true;
         }
 
         setTranslationStringOptions();
@@ -8263,6 +8367,10 @@ function calendarJs( id, options, searchOptions, startDateTime ) {
 
         if ( !isDefinedString( _options.showHolidaysInTheDisplaysText ) ) {
             _options.showHolidaysInTheDisplaysText = "Show holidays in the main display and title bars";
+        }
+
+        if ( !isDefinedString( _options.newEventDefaultTitle ) ) {
+            _options.newEventDefaultTitle = "* New Event";
         }
     }